- ssize_t type fixes
git-svn-id: https://svn.code.sf.net/p/boxi/code/trunk/boxi@473 bde4d9af-c013-0410-a270-fa18cafdf79f

+++ src/TestWithServer.cc
@@ -373,8 +373,8 @@ void TestWithServer::CompareFiles(const wxFileName& first,
 		{
 			toread = a.Length() - pos;
 		}
-		CPPUNIT_ASSERT_EQUAL(toread, a.Read(buffer1, toread));
-		CPPUNIT_ASSERT_EQUAL(toread, b.Read(buffer2, toread));
+		CPPUNIT_ASSERT_EQUAL((ssize_t)toread, a.Read(buffer1, toread));
+		CPPUNIT_ASSERT_EQUAL((ssize_t)toread, b.Read(buffer2, toread));
 		CPPUNIT_ASSERT_EQUAL(0, memcmp(buffer1, buffer2, toread));
 	}
 }
